Revert "very basic profiles without attempting to make anything pretty (#242)"
This reverts commit e07dc116b366b5ba186fd5c5a15ffeee46b5425c.
