+++ frontend/src/store/index.js
@@ -5,7 +5,6 @@ import reviews from './reviews'
 import course from './course'
 import auth from './auth'
 import subject from './subject'
-import user from './user'
 import global from './root'
 
 Vue.use(Vuex)
@@ -16,8 +15,7 @@ export default new Vuex.Store({
     questions,
     reviews,
     auth,
-    subject,
-    user
+    subject
   },
   ...global
 })
