--- frontend/src/utils/api/user.js
@@ -1,6 +1,0 @@
-import { get } from './index'
-
-/* get user's public profile */
-export function getUser(id) {
-  return get(`/user/${id}`)
-}
